requirements.txt
PyCogent>=1.3.1
matplotlib=0.98.5.2-1
numpy>=1.3.0
